Задание 7. Изименил калькулятор чтобы применить и проверить работу исключений. Вместо типа double теперь используется int в вычислениях чтобы они возникали, например, при делении на 0, или вводе не целых чисел.

+++ src/main/java/Five/CalculatorOperation/Minus.java
@@ -2,7 +2,7 @@ package Five.CalculatorOperation;
 
 public class Minus extends Operation {
 
-    public Minus(double x, double y) {
+    public Minus(int x, int y) {
         super(x, y);
         System.out.println(x - y);
