+++ src/main/java/Five/CalculatorOperation/Multiplication.java
@@ -2,7 +2,7 @@ package Five.CalculatorOperation;
 
 public class Multiplication extends Operation {
 
-    public Multiplication(double x, double y) {
+    public Multiplication(int x, int y) {
         super(x, y);
         System.out.println(x * y);
     }
