+++ src/main/java/Five/CalculatorOperation/Plus.java
@@ -2,7 +2,7 @@ package Five.CalculatorOperation;
 
 public class Plus extends Operation {
 
-    public Plus(double x, double y) {
+    public Plus(int x, int y) {
         super(x, y);
         System.out.println(x + y);
     }
